Including rules for quotes and commas in base and react

+++ base.js
@@ -16,7 +16,7 @@ module.exports = {
     rules: {
         quotes: [
             'warn',
-            'double',
+            'single',
             { avoidEscape: true, allowTemplateLiterals: true },
         ],
         'comma-dangle': ['error', 'always'],
